pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Dmitry Shevchenko <dmitry.e.shev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
motor = "^3.0.0"
beanie = "^1.11.6"
bcrypt = "^3.2.2"
fastapi = "^0.78.0"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
testfixtures = "^6.18.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
